requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
transformers==4.38.2
torch==2.2.0
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3 